requirements.txt
fastapi
uvicorn
websockets
deepgram-sdk
langchain
langchain-groq
langchain-core
edge-tts
python-dotenv